tdb-console/requirements/development.txt
aliyun-python-sdk-core==2.3.1
aliyun-python-sdk-sts==2.1.6
appdirs==1.4.3
click==6.6
crcmod==1.7
dnspython==1.15.0
ecdsa==0.13
et-xmlfile==1.0.1
etcd==2.0.8
Flask==0.11.1
Flask-Excel==0.0.5
Flask-WTF==0.12
itsdangerous==0.24
jdcal==1.3
Jinja2==2.8
MarkupSafe==0.23
openpyxl==2.4.0
oss2==2.2.0
packaging==16.8
paramiko==1.17.2
Pillow==3.3.1
pycrypto==2.6.1
pyexcel==0.2.5
pyexcel-io==0.2.3
pyexcel-webio==0.0.7
supervisor==3.3.1
pyexcel-xls==0.2.3
pyparsing==2.2.0
python-etcd==0.4.3
pytz==2013.8
requests==2.12.3
six==1.10.0
texttable==0.8.4
urllib3==1.19.1
Werkzeug==0.11.11
WTForms==2.1
xlrd==1.0.0
xlwt==1.1.2
